Progress on testing enable/disable
