--- SDK/AppCenterCrashes/Microsoft.AppCenter.Crashes.Windows.Shared/Utils/IStorageOperationsHelper.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AppCenter.Crashes.Ingestion.Models;
-
-namespace Microsoft.AppCenter.Crashes.Utils
-{
-    public interface IStorageOperationsHelper
-    {
-        void SaveErrorLogFile(ManagedErrorLog errorLog);
-    }
-}
