+++ SDK/AppCenterCrashes/Microsoft.AppCenter.Crashes.WindowsDesktop/Microsoft.AppCenter.Crashes.WindowsDesktop.csproj
@@ -48,7 +48,6 @@
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils\ProcessInformation.cs" />
-    <Compile Include="Utils\StorageOperationsHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\AppCenter\Microsoft.AppCenter.WindowsDesktop\Microsoft.AppCenter.WindowsDesktop.csproj">
